remove SwiftLayoutPrinter from SwiftLayout add SwiftLayoutPrinter library
